+++ index.html
@@ -233,5 +233,6 @@
     </div>
   </template>
 
+  <script scr="js/main.js" type="'module"></script>
 </body>
 </html>
